menuswap-scraper/pyproject.toml
[project]
name = "menuswap-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiodns==3.2.0",
    "aiohttp==3.9.5",
    "async-timeout==4.0.3",
    "beautifulsoup4==4.12.3",
    "lxml==5.2.2",
    "psycopg[binary]==3.2.1",
    "pydantic==2.8.2",
    "python-dotenv==1.0.1",
    "rapidfuzz==3.9.1",
    "tqdm==4.66.4",
    "boto3==1.34.162",
]
